Refactor: useWalletState 삭제
walletContext로 합쳐서 사용
