+++ src/components/wallet/AmountWrap.jsx
@@ -1,9 +1,9 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import { Container, Amount } from 'components/wallet/AmountWrap.style';
-import useWalletState from 'hooks/useWalletState';
+import { WalletSetContext } from 'contexts/WalletProvider';
 
 export default function AmountWrap() {
-  const { getAmount } = useWalletState();
+  const { getAmount } = useContext(WalletSetContext);
   return (
     <Container>
       총<Amount>{getAmount()}</Amount>원
